Moved sub-classes of "DataFormats", "DataGridViewComboBoxCell", "DataGridViewHeaderCell" and "CategoryGridEntry" classes to their own files
